Added more tests to new Tabs
